requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-Migrate>=4.0
Flask-WTF>=1.1
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=20.1
passlib[bcrypt]>=1.7
email-validator>=1.3
